Updated Psurface constructor to take a surface instance.
My initial goal of having the Psurface constructor only loosely rely on the
SC.Surface instance (by passing the params) has turned out to be an
impediment to further improvements to Psurface.  In particular, I need the
ability for an SC.Surface instance to customize its DOM node when it is:

a) first created, but not yet added to the DOM,
b) after it has been added to the DOM, and possibly
c) when the DOM node is removed from the DOM.

This is the first patch to enable this.  The psurfaces_test app runs without
errors.