Added SC.TextSurface instance to form_demo, Notes tab (demonstrates bug).
This commit demonstrates a bug.  Click on the Characteristics tab, the back
to the Notes tab.  Observe that the SC.TextSurface is positioned incorrectly.
This occurs because SC.Psurface only applies SC.Surface properties when the
property is animated, not when the DOM element is first created.  The next
commit fixes this.