added SC.View#updateEventForLayer() method and updated mouseDragged:
When mouseDragged: is called during a drag session (SC.app.dragDidStart())
and an event was passed to dragDidStart(), the hitPoint, layerX, and layerY
will be updated using SC.View#updateEventForLayer() prior to invoking
mouseDragged on the drag responder.

You can also call updateEventForLayer() is your own code (see docs) as
needed. Doing so provides an easy way to get the mouse coordinates in the
layer's local coordinate system.